Use values for boolean config options
Currently we enable features if the according config macro is defined
and disable it if it is undefined. This requires to use the preprocessor
to evaluate the config options and leaves code paths that are never fed
into the compiler.

Ideally we can get rid of these code paths and let the compiler handle
the evaluation of the config options. To prepare for this we need to
use values for the boolean config option macros (1 for enabled and 0
for disabled).

No functional changes intended.